Conditionally rendered Add To List button if Anime is on User List
Included a list of ids of animes that are on the users list when retrieving a user. Also made the add to list button render conditionally based on whether the anime is on the logged in users list or not. Created api endpoint for retrieving the users list, and retrieving the UserAnime details.