Reformatted loading logic on anime detail page
Reformatted the loading logic in the anime details page to render the UserAnime details depending on whether or not the anime is on the users list. I still need to make it so that when the user clicks the add to list button, it switched to the user anime inputs.