Added refresh token logic
Made login and register endpoints return a secure cookie containing the refresh token along with the user and access token. Also created a refresh endpoint that generates new access tokens. Used axios request interceptor to check if access token is expired, and if so, sends a refresh request to get a new access token, and then makes the request.