Fixed refresh token logic
Made it so refresh tokens are stored in the database so they can be tracked and invalidated. Upon logging in, the user is granted an access and refresh token and the refresh token is stored in the db. Upon logging out, the refresh token is kept in the db but is invalidated. Also fixed issue where Featured Carousel was not an observer of the user store. Also fixed issue where items on the list page were not taking up 100% of the width of their container.