Added a hover effect to anime cards on home page.
Added a hover effect on the anime cards on the home page that displays their title, description, rating, and number of episodes with an overlay. Also did some reformatting to handle errors for adding and removing anime from list in their userStore actions rather than needing to use a formik form.